+++ samples/kmp-project-template/sync/src/iosMain/kotlin/org/mifos/sync/IosSyncManager.kt
@@ -3,7 +3,8 @@ package org.mifos.sync
 import io.github.mobilebytelabs.worker.WorkManager
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import org.mifos.core.data.util.SyncManager
+import io.github.mobilebytelabs.worker.scheduler.WorkScheduler
+import io.github.mobilebytelabs.worker.scheduler.sync.SyncManager
 
 class IosSyncManager(private val workManager: WorkManager, private val scheduler: WorkScheduler) : SyncManager {
     private val _isSyncing = MutableStateFlow(false)